ML_deplor/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
joblib==1.3.2
gunicorn==21.2.0
torch
opencv-python
google-genai
gTTS
requests
